backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
gunicorn>=21.2,<23.0
whitenoise>=6.6,<7.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
